Fix racy deployment rollback
Summary:
recent instance's goal related refactorings introduced a race condition where decommissioning a _Scheduled_ instance would be unhandled by the `TaskReplaceActor` so that a deployment would be stuck since not enough instance would be launched. This patch fixes this bug. A regression integration test was added. Additionally we get rid of 2/3 `LaunchQueue.purge` calls that existed in and around deployments. These are not needed anymore.

Fixes: MARATHON-8537
Related: MARATHON-8223